refactor(node): split up
