+++ src/server/utils/constants.ts
@@ -5,7 +5,7 @@ import {
   IS_IN_PRODUCTION,
   IS_IN_STACK,
   IS_NITRO_OPENAPI_ENABLED,
-} from '../../node'
+} from '../../node/environment'
 
 export const DARGSTACK_SECRET_UNUSED_THIRD_PARTY = 'UNSET THIRD PARTY SECRET'
 export const GET_CSP = ({
